streamlining mapper for hardware and sim
