+++ jetracer_description/urdf/jetracer_joints.xacro
@@ -31,10 +31,31 @@
         <child link="base_footprint" />
     </joint>
 
-    <joint name="laser_joint" type="fixed">
-        <origin xyz="0.005 0 0.120" rpy="0 0 0" />
-        <parent link="base_link" />
-        <child link="laser" />
+    <xacro:if value="$(arg use_sim)">
+        <joint name="laser_joint" type="fixed">
+            <origin xyz="0.005 0 0.120" rpy="0 0 0" />
+            <parent link="base_link" />
+            <child link="laser" />
+        </joint>
+    </xacro:if>
+    <xacro:unless value="$(arg use_sim)">
+        <joint name="laser_joint" type="fixed">
+            <origin xyz="0.005 0 0.120" rpy="0 0 ${pi}" />
+            <parent link="base_link" />
+            <child link="laser" />
+        </joint>
+    </xacro:unless>
+
+    <joint name="camera_joint" type="fixed">
+        <parent link="Camera_link"/>
+        <child link="camera_link"/>
+        <origin xyz="0.015019559999997335 0.00105 -0.015019555" rpy="-1.5707963267948903 1.4594420859336604e-31 3.8721067230568025e-16"/>
+    </joint>
+
+    <joint name="camera_optical_joint" type="fixed">
+        <parent link="camera_link"/>
+        <child link="camera_link_optical"/>
+        <origin xyz="0 0 0" rpy="${-pi/2} 0 ${-pi/2}"/>
     </joint>
 
     <!-- ================ JOINTS ================ -->
