+++ jetracer_description/urdf/jetracer_links.xacro
@@ -35,6 +35,8 @@
     <link name="base_link" />
     <link name="base_footprint" />
     <link name="laser" />
+    <link name="camera_link"/>
+    <link name="camera_link_optical"/>
 
     <!-- ================ JOINTS ================ -->
     <!-- chassis link -->
