+++ jetracer_description/urdf/jetracer_sensors.xacro
@@ -26,22 +26,6 @@
         </sensor>
     </gazebo>
 
-    <joint name="camera_joint" type="fixed">
-        <parent link="Camera_link"/>
-        <child link="camera_link"/>
-        <origin xyz="0.015019559999997335 0.00105 -0.015019555" rpy="-1.5707963267948903 1.4594420859336604e-31 3.8721067230568025e-16"/>
-    </joint>
-
-    <link name="camera_link"/>
-
-    <joint name="camera_optical_joint" type="fixed">
-        <parent link="camera_link"/>
-        <child link="camera_link_optical"/>
-        <origin xyz="0 0 0" rpy="${-pi/2} 0 ${-pi/2}"/>
-    </joint>
-
-    <link name="camera_link_optical"/>
-
     <gazebo reference="camera_link">
         <material>Gazebo/Red</material>
